requirements.txt
Appium-Python-Client~=1.3.0
selenium~=3.141.0
PyYAML~=6.0
parameterized~=0.8.1
HTMLTestReport~=1.0.1